# perf: fix multiprocessing timing measurement